Move picture attachment save listener to standard long click menu
When long-clicking a message with an image attachment, add a "Save Image" menu item to the context menu.
This context menu also applies to the image, so long-clicking the image opens the messages context menu.
This is particularly useful for picture messages without text.
