+++ res/menu/conversation_context_image.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:title="@string/conversation_context_image__menu_save"
+          android:id="@+id/menu_context_save_image" />
+</menu>
